Instantiate db for each update operation
Before this commit, the db connection was instantiated once and kept
alive throughout the process. However, this led to the program not
terminating because it didn't know when the db writes would end.

This change moves the db instantiation to each function call, trading
off performance for the program's ability to terminate correctly.
[Finishes #118163501]