Keep db connection open for now
An earlier commit moved the db instantiation code into the Upsert
function call, so that the db connection would be closed after each
write. However this caused performance issues and premature closing of
the connections for some writes. Moving the instantiation call back to
the top of the file for now; will address the db.close() issue later.