Implement tag scraping logic for Saks
It's important for us to know the categories assigned to each pair
of shoe by each retailer, so that we can validate the classification
algorithms and support filtering shoes by category in the frontend in
the future.

This first commit implements the scraping logic, and adds support for
tag scraping for Saks. Other retailers to follow soon.
[#118736111]